# api/swagger.yml
...
consumes:
- application/json
produces:
- application/json
swagger: "2.0"
info:
  description: API documentation
  title: github.com/hzbay/chain-bridge
  version: 0.1.0
paths:
...
  /-/tokens/{token_id}:
    get:
      security:
      - Management: []
      description: Retrieves detailed information for a specific token
      tags:
      - cpop mgmt
      summary: Get token by ID
      operationId: getToken
      parameters:
      - type: integer
        description: The token ID
        name: token_id
        in: path
        required: true
      responses:
        "200":
          description: Successfully retrieved token
          schema:
            $ref: '#/definitions/tokenResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Bad request
        "404":
          description: Not found error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Not found
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Internal server error
    put:
      security:
      - Management: []
      description: Updates the configuration for a specific token
      tags:
      - cpop mgmt
      summary: Update token configuration
      operationId: updateToken
      parameters:
      - type: integer
        description: The token ID
        name: token_id
        in: path
        required: true
      - description: Token update request
        name: request
        in: body
        required: true
        schema:
          $ref: '#/definitions/updateTokenRequest'
      responses:
        "200":
          description: Token updated successfully
          schema:
            $ref: '#/definitions/tokenConfigUpdateResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Bad request
        "404":
          description: Not found error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Not found
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Internal server error
    delete:
      security:
      - Management: []
      description: Removes a token from the supported tokens list
      tags:
      - cpop mgmt
      summary: Delete token
      operationId: deleteToken
      parameters:
      - type: integer
        description: The token ID
        name: token_id
        in: path
        required: true
      responses:
        "200":
          description: Token deleted successfully
          schema:
            $ref: '#/definitions/tokenDeleteResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Bad request
        "404":
          description: Not found error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Not found
        "409":
          description: Token is in use and cannot be deleted
          schema:
            $ref: '#/definitions/apiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Internal server error
  /-/tokens/{token_id}/toggle-status:
    put:
      security:
      - Management: []
      description: Enables or disables a specific token
      tags:
      - cpop mgmt
      summary: Toggle token enabled status
      operationId: toggleTokenStatus
      parameters:
      - type: integer
        description: The token ID
        name: token_id
        in: path
        required: true
      - type: boolean
        description: Whether to enable or disable the token
        name: enabled
        in: query
        required: true
      responses:
        "200":
          description: Token status updated successfully
          schema:
            $ref: '#/definitions/tokenConfigUpdateResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Bad request
        "404":
          description: Not found error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Not found
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Internal server error
...
definitions:
...
  apiError:
    type: object
    required:
    - message
    properties:
      errors:
        type: array
        items:
          type: object
          properties:
            code:
              description: Error code
              type: string
            field:
              description: Field name
              type: string
            message:
              description: Error message
              type: string
      message:
        description: Error message
        type: string
        example: Bad request
...
  tokenConfigUpdateResponse:
    type: object
    required:
    - message
    properties:
      message:
        type: string
        example: Token configuration updated successfully
...
  tokenDeleteResponse:
    type: object
    required:
    - message
    properties:
      message:
        type: string
        example: Token deleted successfully
  tokenResponse:
    type: object
    properties:
      batch_operations:
        description: Supported batch operations configuration
        type: object
        additionalProperties: true
        example:
          batchBurn: true
          batchMint: true
          batchTransfer: true
      chain_id:
        description: Chain ID
        type: integer
        format: int64
        example: 56
      contract_address:
        description: Token contract address (null for native tokens)
        type: string
        example: 0x1234567890123456789012345678901234567890
      created_at:
        description: Token creation timestamp
        type: string
        format: date-time
        example: "2024-12-21T00:00:00Z"
      decimals:
        description: Token decimals
        type: integer
        example: 18
      id:
        description: Token ID
        type: integer
        example: 1
      is_enabled:
        description: Whether token is enabled
        type: boolean
        example: true
      name:
        description: Token name
        type: string
        example: CPOP Token
      supports_batch_operations:
        description: Whether token supports batch operations
        type: boolean
        example: true
      symbol:
        description: Token symbol
        type: string
        example: CPOP
      token_type:
        description: Token type
        type: string
        enum:
        - native
        - erc20
        example: erc20
...
  updateTokenRequest:
    type: object
    properties:
      batch_operations:
        description: Supported batch operations configuration
        type: object
        additionalProperties: true
        example:
          batchBurn: false
          batchMint: true
          batchTransfer: true
      is_enabled:
        description: Whether token is enabled
        type: boolean
        example: false
      name:
        description: Token name
        type: string
        example: Updated Token Name
      supports_batch_operations:
        description: Whether token supports batch operations
        type: boolean
        example: true
...
securityDefinitions:
  Bearer:
    description: |-
      Access token for application access, **must** include "Bearer " prefix.
      Example: `Bearer b4a94a42-3ea2-4af3-9699-8bcbfee6e6d2`
    type: apiKey
    name: Authorization
    in: header
    x-keyPrefix: 'Bearer '
  Management:
    description: Management secret, used for monitoring and infrastructure related
      calls
    type: apiKey
    name: mgmt-secret
    in: query